requirements.txt
streamlit==1.30.0
pandas==1.5.3
joblib==1.4.2
shap==0.45.1
numpy==1.24.3
pillow==10.0.0
matplotlib==3.10.0
scikit-learn==1.5.1